okschooldata: Fetch and Process Oklahoma School Data

Downloads and processes school data from the Oklahoma State Department of Education (OSDE). Provides functions for fetching enrollment data from OSDE's public reporting systems and transforming it into tidy format for analysis.

Main functions

fetch_enr

Fetch enrollment data for a school year

fetch_enr_multi

Fetch enrollment data for multiple years

get_available_years

Get list of available data years

tidy_enr

Transform wide data to tidy (long) format

id_enr_aggs

Add aggregation level flags

enr_grade_aggs

Create grade-level aggregations

Cache functions

cache_status

View cached data files

clear_cache

Remove cached data files

ID System

Oklahoma uses an alphanumeric ID system:

District IDs: County code (2 digits) + Type (1 letter) + Number (3 digits)

Site IDs: District ID + Site number (3 digits)

Example: 55I001 = Oklahoma County (55), Independent (I), District 001 (OKC Public Schools)

Example: 55I001001 = Site 001 in Oklahoma City Public Schools

District Types

I = Independent school district

D = Dependent school district

C = City school district

E = Elementary school district

Data Sources

Data is sourced from the Oklahoma State Department of Education:

OSDE Public Records: https://sde.ok.gov/reporting-index

State Public Enrollment: https://sde.ok.gov/documents/state-student-public-enrollment

OklaSchools.com: https://oklaschools.com/

See also

Author

Maintainer: Al Martin almartin@example.com